requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
contentful==1.12.2
Flask==1.1.1
Flask-Markdown==0.3
Frozen-Flask==0.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
pbr==5.4.1
Pillow==6.1.0
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
stevedore==1.30.1
urllib3==1.25.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.15.5
